abstract out execute
