+++ src/flowCell/flowCell.js
@@ -1,13 +1,10 @@
-import { formatElapsedTime } from '../utils/formatElapsedTime';
-import { formatClockTime } from '../utils/formatClockTime';
+/* eslint no-unused-vars: "error"*/
+
 import templateOf from './templateOf';
 import scrollIntoView from './scrollIntoView';
 import autoResize from './autoResize';
 import getCursorPosition from './getCursorPosition';
-import endFunction from './endFunction';
-import errorFunction from './errorFunction';
-import closeFunction from './closeFunction';
-import dataFunction from './dataFunction';
+import execute from './execute';
 
 export function flowCell(_, type, input) {
   console.log('arguments from flowCell', arguments);
@@ -94,54 +91,6 @@ export function flowCell(_, type, input) {
       return _hasInput(true);
     }
   };
-  const execute = go => {
-    const startTime = Date.now();
-    _time(`Started at ${formatClockTime(startTime)}`);
-    input = _input().trim();
-    if (!input) {
-      if (go) {
-        return go(null);
-      }
-      return void 0;
-    }
-    const render = _render();
-    _isBusy(true);
-    clear();
-    if (_type() === 'sca') {
-      // escape backslashes
-      input = input.replace(/\\/g, '\\\\');
-      // escape quotes
-      input = input.replace(/'/g, '\\\'');
-      // escape new-lines
-      input = input.replace(/\n/g, '\\n');
-      // pass the cell body as an argument, representing the scala code, to the appropriate function
-      input = `runScalaCode ${_.scalaIntpId()}, \'${input}\'`;
-    }
-    const outputObject = {
-      data: dataFunction.bind(this, _outputs),
-      close: closeFunction.bind(this, _result),
-      error: errorFunction.bind(
-        this,
-        _,
-        _hasError,
-        _outputs,
-        _errors
-      ),
-      end: endFunction.bind(
-        this,
-        _hasInput,
-        _isCode,
-        _isBusy,
-        _time,
-        _hasError,
-        _errors,
-        startTime,
-        go
-      ),
-    };
-    render(input, outputObject);
-    return _isActive(false);
-  };
   const self = {
     guid: _guid,
     type: _type,
@@ -164,7 +113,24 @@ export function flowCell(_, type, input) {
     select,
     navigate,
     activate,
-    execute,
+    execute: execute.bind(
+      this,
+      _,
+      _time,
+      input,
+      _input,
+      _render,
+      _isBusy,
+      clear,
+      _type,
+      _outputs,
+      _result,
+      _hasError,
+      _errors,
+      _hasInput,
+      _isActive,
+      _isCode
+    ),
     clear,
     clip,
     _actions,
